Split check graph validation fixtures
