+++ docs/COMPLETION_AUDIT.md
@@ -2996,6 +2996,11 @@ and do not assume Phase 4 is ready without evidence.
   `tests/integration/cli_build/direct_build_graph_execution.rs` from 417 to
   198 lines while preserving coverage through
   `cargo test --test integration direct_file_command_build_zen`.
+- Check-command build.zen integration tests now keep graph validation cases in
+  `tests/integration/cli_build/graph_validation/`, split across basic
+  acceptance, dependency-shape rejection, and source-validation modules while
+  preserving coverage through
+  `cargo test --test integration graph_validation`.
 - Emit build.zen integration tests now keep graph validation cases in
   `tests/integration/cli_build/emit_direct_validation.rs`, reducing
   `tests/integration/cli_build/emit_direct.rs` from 377 to 47 lines while
