+++ tests/integration/cli_build/graph_validation/mod.rs
@@ -1,0 +1,3 @@
+mod basics;
+mod dependency_shapes;
+mod source_validation;
